Feat: Auth 관련 Service 인터페이스 생성 (#12)
